Add 3D plotting functionality and related commands
- Introduced a new module `plot3d.py` for custom Gkeyll 3D plotting using Plotly.
- Implemented the `plot3d` function to visualize 3D data with various options for styling, axis scaling, and color mapping.
- Added a command `animate3d` to the CLI for generating animated 3D plots.
- Updated the main CLI file `pgkyl.py` to include the new `animate3d` and `plot3d` commands.
