+++ src/postgkyl/commands/plot.py
@@ -1,66 +1,10 @@
 import click
-import importlib
 import matplotlib.pyplot as plt
 import numpy as np
-import os.path
-from pathlib import Path
-import webbrowser
-from postgkyl.data import GData
-from postgkyl.data.select import select as data_select
 
 from postgkyl.utils import verb_print
+import postgkyl.output.plot
 
-
-def _parse_range_option(_ctx, _param, value):
-  if value is None:
-    return None
-  # end
-
-
-def _parse_slice_option(_ctx, _param, value):
-  if value is None:
-    return None
-  # end
-
-  tokens = [token.strip() for token in str(value).split(",") if token.strip()]
-  if not tokens:
-    raise click.BadParameter("Expected a number or comma-separated list of numbers.")
-  # end
-
-  selectors = []
-  for token in tokens:
-    token_lower = token.lower()
-    # Int tokens are interpreted as indices, float tokens as coordinates.
-    if "." in token_lower or "e" in token_lower:
-      try:
-        selectors.append(float(token))
-      except ValueError as exc:
-        raise click.BadParameter(
-            f"Invalid selector '{token}'. Use int for index or float for coordinate value."
-        ) from exc
-      # end
-    else:
-      try:
-        selectors.append(int(token))
-      except ValueError as exc:
-        raise click.BadParameter(
-            f"Invalid selector '{token}'. Use int for index or float for coordinate value."
-        ) from exc
-      # end
-    # end
-  # end
-  return selectors
-
-  parts = [part.strip() for part in value.replace(":", ",").split(",") if part.strip()]
-  if len(parts) != 2:
-    raise click.BadParameter("Expected two numbers in the form 'lower,upper' or 'lower:upper'.")
-  # end
-
-  try:
-    return (float(parts[0]), float(parts[1]))
-  except ValueError as exc:
-    raise click.BadParameter("Expected two numbers in the form 'lower,upper' or 'lower:upper'.") from exc
-  # end
 
 @click.command()
 @click.option("--use", "-u", default=None, help="Specify the tag to plot.")
@@ -88,69 +32,40 @@ def _parse_slice_option(_ctx, _param, value):
 @click.option("--linewidth", type=click.FLOAT, help="Set the linewidth.")
 @click.option("--linestyle", type=click.Choice(["solid", "dashed", "dotted", "dashdot"]),
     help="Set the linestyle.")
-@click.option("-o","--opacity", type=click.FLOAT, default=1.0, help="Set opacity for 3D volume plots (0.0-1.0).")
-@click.option("--surface-count", type=click.INT, default=32, show_default=True,
-  help="Number of Plotly volume isosurfaces to render for 3D plots.")
-@click.option("--maximum-points-per-axis", "--mppa", "maximum_points_per_axis", type=click.INT, default=0, show_default=True,
-  help="Maximum number of points along any 3D volume axis; 0 disables downsampling.")
 @click.option("--style", help="Specify Matplotlib style file (default: Postgkyl).")
-@click.option("--background", type=click.Choice(["dark", "light"]), default="dark", show_default=True,
-    help="Background mode for plots (dark/light).")
 @click.option("-d", "--diverging", is_flag=True, help="Switch to diverging color map.")
 @click.option("--arg", type=click.STRING, default="",
     help="Additional plotting arguments, e.g., '*--'.")
 @click.option("--fix-aspect", "-a", "fixaspect", is_flag=True,
     help="Enforce the same scaling on both axes.")
-@click.option("--aspect", default=None,
-  help="Specify aspect behavior. For Plotly 3D use one of: auto,data,cube (or a numeric ratio).")
+@click.option("--aspect", default=None, help="Specify the scaling ratio.")
 @click.option("--logx", is_flag=True, help="Set x-axis to log scale.")
 @click.option("--logy", is_flag=True, help="Set y-axis to log scale.")
-@click.option("--logz", is_flag=True, help="Set z-axis (in 2D, values of the plot) to log scale.")
-@click.option("--logc", is_flag=True, help="Set colorbar to log scale for 3D plots.")
+@click.option("--logz", is_flag=True, help="Set values of 2D plot to log scale.")
 @click.option("--xshift", default=0.0, type=click.FLOAT, show_default=True,
     help="Value to shift the x-axis.")
 @click.option("--yshift", default=0.0, type=click.FLOAT, show_default=True,
     help="Value to shift the y-axis.")
 @click.option("--zshift", default=0.0, type=click.FLOAT, show_default=True,
     help="Value to shift the z-axis.")
-@click.option("--cshift", default=0.0, type=click.FLOAT, show_default=True,
-    help="Value to shift the color values for 3D plots.")
 @click.option("--xscale", default=1.0, type=click.FLOAT, show_default=True,
     help="Value to scale the x-axis.")
 @click.option("--yscale", default=1.0, type=click.FLOAT, show_default=True,
     help="Value to scale the y-axis.")
 @click.option("--zscale", default=1.0, type=click.FLOAT, show_default=True,
     help="Value to scale the z-axis (default: 1.0).")
-@click.option("--slice-at-z0", type=click.STRING, callback=_parse_slice_option, default=None,
-  help="Select z0 slices. Comma-separated selectors; ints are indices, floats are coordinate values.")
-@click.option("--slice-at-z1", type=click.STRING, callback=_parse_slice_option, default=None,
-  help="Select z1 slices. Comma-separated selectors; ints are indices, floats are coordinate values.")
-@click.option("--slice-at-z2", type=click.STRING, callback=_parse_slice_option, default=None,
-  help="Select z2 slices. Comma-separated selectors; ints are indices, floats are coordinate values.")
-@click.option("--slice-at-z3", type=click.STRING, callback=_parse_slice_option, default=None,
-  help="Select z3 slices. Comma-separated selectors; ints are indices, floats are coordinate values.")
-@click.option("--slice-at-z4", type=click.STRING, callback=_parse_slice_option, default=None,
-  help="Select z4 slices. Comma-separated selectors; ints are indices, floats are coordinate values.")
-@click.option("--slice-at-z5", type=click.STRING, callback=_parse_slice_option, default=None,
-  help="Select z5 slices. Comma-separated selectors; ints are indices, floats are coordinate values.")
-@click.option("--cscale", default=1.0, type=click.FLOAT, show_default=True,
-    help="Value to scale the color values for 3D plots.")
 @click.option("--xmax", default=None, type=click.FLOAT, help="Set maximal x-value.")
 @click.option("--xmin", default=None, type=click.FLOAT, help="Set minimal x-values.")
 @click.option("--ymax", default=None, type=click.FLOAT, help="Set maximal y-value.")
 @click.option("--ymin", default=None, type=click.FLOAT, help="Set minimal y-values.")
 @click.option("--zmax", default=None, type=click.FLOAT, help="Set maximal z-value.")
 @click.option("--zmin", default=None, type=click.FLOAT, help="Set minimal z-values.")
-@click.option("--cmax", default=None, type=click.FLOAT, help="Set maximal color value for 3D plots.")
-@click.option("--cmin", default=None, type=click.FLOAT, help="Set minimal color value for 3D plots.")
-@click.option("--xlim", default=None, type=click.STRING, callback=_parse_range_option,
+@click.option("--xlim", default=None, type=click.STRING,
     help="Set limits for the x-coordinate (lower,upper)")
-@click.option("--ylim", default=None, type=click.STRING, callback=_parse_range_option,
+@click.option("--ylim", default=None, type=click.STRING,
     help="Set limits for the y-coordinate (lower,upper).")
-@click.option("--zlim", default=None, type=click.STRING, callback=_parse_range_option,
+@click.option("--zlim", default=None, type=click.STRING,
     help="Set limits for the z-coordinate (lower,upper).")
-@click.option("--clim", default=None, type=click.STRING, callback=_parse_range_option,
-    help="Set limits for the color scale (lower,upper).")
 @click.option("--relax", is_flag=True, help="Relax the stringent x axis limits for 1D plots.")
 @click.option("--globalrange", "-r", is_flag=True, help="Make uniform extends across datasets.")
 @click.option("--cutoffglobalrange", "-cogr", default=None, type=click.FLOAT,
@@ -163,7 +78,6 @@ def _parse_slice_option(_ctx, _param, value):
 @click.option("--color", type=click.STRING, help="Set color when available.")
 @click.option("-x", "--xlabel", type=click.STRING, help="Specify a x-axis label.")
 @click.option("-y", "--ylabel", type=click.STRING, help="Specify a y-axis label.")
-@click.option("-z", "--zlabel", type=click.STRING, help="Specify a z-axis label.")
 @click.option("--clabel", type=click.STRING, help="Specify a label for colorbar.")
 @click.option("--title", type=click.STRING, help="Specify a title.")
 @click.option("--subplot-titles", type=click.STRING, help="Comma-separated titles for each subplot. e.g. --subplot-titles 'Title1,Title2,Title3'")
@@ -171,15 +85,6 @@ def _parse_slice_option(_ctx, _param, value):
 @click.option("--subplot-ylabels", type=click.STRING, help="Comma-separated y-axis labels for each subplot. e.g. --subplot-ylabels 'Y1,Y2,Y3'")
 @click.option("--save", is_flag=True, help="Save figure as PNG file.")
 @click.option("--saveas", type=click.STRING, default=None, help="Name of figure file.")
-@click.option("--starting-azimuthal-angle", "azimuthal_angle", "--azimuthal-angle",
-  type=click.FLOAT, default=0.0, show_default=True,
-  help="Starting azimuthal angle in degrees for rotating 3D save.")
-@click.option("--polar-angle", type=click.FLOAT, default=85.0, show_default=True,
-  help="Polar angle in degrees for rotating 3D camera. 90 degrees is the x-y plane.")
-@click.option("--rotation-period", type=click.FLOAT, default=20.0, show_default=True,
-  help="Rotation period in seconds for one full rotation (used for rotating html/mp4/gif output).")
-@click.option("--fps", type=click.INT, default=1, show_default=True,
-  help="FPS used for rotating mp4/gif save output.")
 @click.option("--dpi", type=click.INT, default=200, help="DPI (resolution) for output.")
 @click.option("-e", "--edgecolors", type=click.STRING,
     help="Set color for cell edges to show grid outline.")
@@ -194,8 +99,6 @@ def _parse_slice_option(_ctx, _param, value):
 @click.option("--jet", is_flag=True, help="Turn colormap to jet for comparison with literature.")
 @click.option("--cmap", type=click.STRING, default=None,
     help="Override default colormap with a valid matplotlib cmap.")
-@click.option("--invert-cmap", is_flag=True,
-  help="Invert the selected colormap (or the default colormap for the chosen background mode).")
 @click.option("-m", "--multiblock", is_flag=True, default=False)
 @click.pass_context
 def plot(ctx, **kwargs):
@@ -204,49 +107,6 @@ def plot(ctx, **kwargs):
   Plot labels can use a sub-set of LaTeX math commands placed between dollar ($) signs.
   """
   verb_print(ctx, "Starting plot")
-  plot_output_module = importlib.import_module("postgkyl.output.plot")
-
-  def _save_output(file_name):
-    plt.savefig(file_name, dpi=kwargs["dpi"])
-
-  def _save_output_3d(fig, file_name: str | None = None, base_name: str | None = None,
-      force_rotating_preview: bool = False) -> str:
-    if force_rotating_preview:
-      safe_base = "".join(ch if ch.isalnum() or ch in ("-", "_") else "_" for ch in (base_name or "")).strip("_")
-      if not safe_base:
-        safe_base = "plot_preview"
-      # end
-      file_name = os.path.join(os.getcwd(), f"{safe_base}_preview.html")
-    elif file_name is None:
-      raise click.ClickException("Internal error: missing output file name for 3D save.")
-    # end
-
-    root, ext = os.path.splitext(file_name)
-    ext = ext.lower()
-    rotating_target = force_rotating_preview or ext in (".mp4", ".gif", ".html")
-    if rotating_target:
-      if ext == "":
-        file_name = f"{file_name}.mp4"
-      # end
-      plot_output_module.save_rotating_plotly_figure(
-          fig,
-          file_name,
-          starting_azimuthal_angle=kwargs["azimuthal_angle"],
-          polar_angle=kwargs["polar_angle"],
-          rotation_period=kwargs["rotation_period"],
-          fps=kwargs["fps"],
-      )
-      return file_name
-    # end
-
-    if ext != ".html":
-      file_name = f"{root}.html" if root else f"{file_name}.html"
-    # end
-    fig.write_html(file_name)
-    return file_name
-
-  def _open_html_preview(html_name: str):
-    webbrowser.open(Path(html_name).resolve().as_uri())
 
   kwargs["rcParams"] = ctx.obj["rcParams"]
 
@@ -271,62 +131,6 @@ def plot(ctx, **kwargs):
     kwargs["lineouts"] = int(kwargs["lineouts"])
   # end
 
-  slice_kwargs = {}
-  for d in range(6):
-    slice_selectors = kwargs.pop(f"slice_at_z{d}")
-    if slice_selectors is not None:
-      slice_kwargs[f"z{d}"] = slice_selectors
-    # end
-  # end
-
-  def _get_slice_kwargs_for_data(dat, allow_multiple_per_axis: bool):
-    if not slice_kwargs:
-      return {}
-    # end
-
-    num_dims = dat.get_num_dims()
-    resolved = {}
-    for key, selectors in slice_kwargs.items():
-      axis = int(key[1:])
-      if axis >= num_dims:
-        raise click.ClickException(
-            f"Cannot use --slice-at-{key} on a {num_dims:d}D dataset."
-        )
-      # end
-      if not selectors:
-        continue
-      # end
-      if allow_multiple_per_axis:
-        resolved[key] = selectors
-      else:
-        if len(selectors) != 1:
-          raise click.ClickException(
-              f"--slice-at-{key} accepts multiple selectors only for 3D plane overlay plots."
-          )
-        # end
-        resolved[key] = selectors[0]
-      # end
-    # end
-    return resolved
-
-  def _get_plot_data(dat):
-    resolved_slice_kwargs = _get_slice_kwargs_for_data(dat, allow_multiple_per_axis=False)
-    if not resolved_slice_kwargs:
-      return dat
-    # end
-
-    selected_grid, selected_values = data_select(dat, comp=None, **resolved_slice_kwargs)
-    selected_dat = GData(
-      file_name=dat._file_name,
-      tag=dat.get_tag(),
-      label=dat.get_custom_label(),
-      ctx=dat.ctx,
-      comp_grid=ctx.obj["compgrid"],
-      load=False,
-    )
-    selected_dat.push(selected_grid, selected_values)
-    return selected_dat
-
   kwargs["num_axes"] = None
   if kwargs["subplots"]:
     kwargs["num_axes"] = 0
@@ -340,25 +144,16 @@ def plot(ctx, **kwargs):
   # end
 
   if kwargs["xlim"]:
-    kwargs["xmin"], kwargs["xmax"] = kwargs["xlim"]
-    kwargs["xrange"] = kwargs["xlim"]
+    kwargs["xmin"] = float(kwargs["xlim"].split(",")[0])
+    kwargs["xmax"] = float(kwargs["xlim"].split(",")[1])
   # end
   if kwargs["ylim"]:
-    kwargs["ymin"], kwargs["ymax"] = kwargs["ylim"]
-    kwargs["yrange"] = kwargs["ylim"]
+    kwargs["ymin"] = float(kwargs["ylim"].split(",")[0])
+    kwargs["ymax"] = float(kwargs["ylim"].split(",")[1])
   # end
   if kwargs["zlim"]:
-    kwargs["zrange"] = kwargs["zlim"]
-    kwargs["zmin"], kwargs["zmax"] = kwargs["zlim"]
-  # end
-  if kwargs["clim"]:
-    kwargs["cmin"], kwargs["cmax"] = kwargs["clim"]
-  # end
-  if kwargs["cmin"] is not None:
-    kwargs["zmin"] = kwargs["cmin"]
-  # end
-  if kwargs["cmax"] is not None:
-    kwargs["zmax"] = kwargs["cmax"]
+    kwargs["zmin"] = float(kwargs["zlim"].split(",")[0])
+    kwargs["zmax"] = float(kwargs["zlim"].split(",")[1])
   # end
 
   dataset_fignum = False
@@ -381,8 +176,7 @@ def plot(ctx, **kwargs):
     vmax = float("-inf")
     v_extrema = np.array([])
     for dat in ctx.obj["data"].iterator(kwargs["use"]):
-      plot_data = _get_plot_data(dat)
-      val = plot_data.get_values() * kwargs["zscale"]
+      val = dat.get_values() * kwargs["zscale"]
       if vmin > np.nanmin(val):
         vmin = np.nanmin(val)
       # end
@@ -429,7 +223,6 @@ def plot(ctx, **kwargs):
   del kwargs["no_legend"]
 
   file_name = ""
-  last_saved_output: str | None = None
 
   # ---- Loop over all the datasets ----
   for i, dat in ctx.obj["data"].iterator(kwargs["use"], enum=True):
@@ -451,27 +244,7 @@ def plot(ctx, **kwargs):
     # end
 
     # ---- Plot ----
-    plot_kwargs = dict(kwargs)
-    if slice_kwargs and dat.get_num_dims() == 3:
-      plot_data = dat
-      plot_kwargs["slice_plane"] = _get_slice_kwargs_for_data(dat, allow_multiple_per_axis=True)
-    else:
-      plot_data = _get_plot_data(dat)
-    # end
-
-    fig = plot_output_module.plot(plot_data, args, label_prefix=label, **plot_kwargs)
-
-    if hasattr(fig, "write_html") and not (kwargs["save"] or kwargs["saveas"]):
-      if dat._file_name:
-        base_name = dat._file_name.split(".")[0]
-      else:
-        base_name = f"plot_{i}"
-      # end
-      html_name = _save_output_3d(fig, base_name=base_name, force_rotating_preview=True)
-      _open_html_preview(html_name)
-      kwargs["show"] = False
-      continue
-    # end
+    postgkyl.output.plot(dat, args, label_prefix=label, **kwargs)
 
     if kwargs["subplots"]:
       kwargs["start_axes"] = kwargs["start_axes"] + dat.get_num_comps()
@@ -490,70 +263,35 @@ def plot(ctx, **kwargs):
           file_name = file_name + "ev_" + ctx.obj["labels"][i].replace(" ", "_")
         # end
       # end
-      if kwargs["figure"] is None:
-        if hasattr(fig, "write_html"):
-          file_name = _save_output_3d(fig, file_name)
-          last_saved_output = file_name
-        else:
-          _save_output(file_name)
-          last_saved_output = file_name
-        # end
-        file_name = ""
-      # end
+    # end
+    if (kwargs["save"] or kwargs["saveas"]) and kwargs["figure"] is None:
+      file_name = str(file_name)
+      plt.savefig(file_name, dpi=kwargs["dpi"])
+      file_name = ""
     # end
 
     if kwargs["saveframes"]:
       file_name = f"{kwargs['saveframes']:s}_{i:d}.png"
-      if hasattr(fig, "write_html"):
-        last_saved_output = _save_output_3d(fig, file_name)
-      else:
-        _save_output(file_name)
-        last_saved_output = file_name
-      # end
+      plt.savefig(file_name, dpi=kwargs["dpi"])
       kwargs["show"] = False
     # end
 
     if "batch_mode" in ctx.obj:
       if ctx.obj["batch_mode"]:
         file_name = f"{ctx.obj['saveframes_prefix']:s}_{i:d}.png"
-        if hasattr(fig, "write_html"):
-          last_saved_output = _save_output_3d(fig, file_name)
-        else:
-          _save_output(file_name)
-          last_saved_output = file_name
-        # end
+        plt.savefig(file_name, dpi=kwargs["dpi"])
         kwargs["show"] = False
       # end
     # end
 
 
   # end
-  if (kwargs["save"] or kwargs["saveas"]) and file_name != "":
+  if (kwargs["save"] or kwargs["saveas"]):
     file_name = str(file_name)
-    if hasattr(fig, "write_html"):
-      last_saved_output = _save_output_3d(fig, file_name)
-    else:
-      _save_output(file_name)
-      last_saved_output = file_name
-    # end
+    plt.savefig(file_name, dpi=kwargs["dpi"])
   # end
 
   if kwargs["show"]:
-    if hasattr(fig, "show") and hasattr(fig, "to_html"):
-      # If a save target already exists, open it directly and avoid creating extra preview files.
-      if kwargs.get("saveas") and last_saved_output and os.path.exists(last_saved_output):
-        _open_html_preview(last_saved_output)
-      elif 'dat' in locals() and getattr(dat, "_file_name", None):
-        preview_base = dat._file_name.split(".")[0]
-        html_name = _save_output_3d(fig, base_name=preview_base, force_rotating_preview=True)
-        _open_html_preview(html_name)
-      else:
-        preview_base = "plot"
-        html_name = _save_output_3d(fig, base_name=preview_base, force_rotating_preview=True)
-        _open_html_preview(html_name)
-      # end
-    else:
-      plt.show()
-    # end
+    plt.show()
   # end
   verb_print(ctx, "Finishing plot")
