+++ src/postgkyl/pgkyl.py
@@ -138,6 +138,7 @@ cli.add_command(cmd.activate)
 cli.add_command(cmd.agyro)
 cli.add_command(cmd.mom_agyro)
 cli.add_command(cmd.animate)
+cli.add_command(cmd.animate3d)
 cli.add_command(cmd.collect)
 cli.add_command(cmd.current)
 cli.add_command(cmd.deactivate)
@@ -163,6 +164,7 @@ cli.add_command(cmd.mask)
 cli.add_command(cmd.gk_energy_balance)
 cli.add_command(cmd.gk_particle_balance)
 cli.add_command(cmd.plot)
+cli.add_command(cmd.plot3d)
 cli.add_command(cmd.pr)
 cli.add_command(cmd.relchange)
 cli.add_command(cmd.select)
